fix(engine-dom): build the inline renderer factory (#3058)
